make ConfigChangeEvent immutable
