+++ dubbo-cluster/src/main/java/org/apache/dubbo/rpc/cluster/router/condition/config/ListenableRouter.java
@@ -24,6 +24,7 @@ import org.apache.dubbo.common.utils.StringUtils;
 import org.apache.dubbo.configcenter.ConfigChangeEvent;
 import org.apache.dubbo.configcenter.ConfigChangeType;
 import org.apache.dubbo.configcenter.ConfigurationListener;
+import org.apache.dubbo.configcenter.DefaultConfigChangeEvent;
 import org.apache.dubbo.configcenter.DynamicConfiguration;
 import org.apache.dubbo.rpc.Invocation;
 import org.apache.dubbo.rpc.Invoker;
@@ -121,7 +122,7 @@ public abstract class ListenableRouter extends AbstractRouter implements Configu
         configuration.addListener(routerKey, this);
         String rule = configuration.getConfig(routerKey);
         if (rule != null) {
-            this.process(new ConfigChangeEvent(routerKey, rule));
+            this.process(new DefaultConfigChangeEvent(routerKey, rule));
         }
     }
 }
