+++ dubbo-configcenter/dubbo-configcenter-api/src/main/java/org/apache/dubbo/configcenter/DefaultConfigChangeEvent.java
@@ -1,0 +1,63 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.dubbo.configcenter;
+
+/**
+ * Default implementation of Config change event.
+ *
+ * @see ConfigChangeType
+ */
+public class DefaultConfigChangeEvent implements ConfigChangeEvent {
+    private String key;
+
+    private String value;
+    private ConfigChangeType changeType;
+
+    public DefaultConfigChangeEvent(String key, String value) {
+        this(key, value, ConfigChangeType.MODIFIED);
+    }
+
+    public DefaultConfigChangeEvent(String key, String value, ConfigChangeType changeType) {
+        this.key = key;
+        this.value = value;
+        this.changeType = changeType;
+    }
+
+    public String getKey() {
+        return key;
+    }
+
+    public void setKey(String key) {
+        this.key = key;
+    }
+
+    public String getValue() {
+        return value;
+    }
+
+    public void setValue(String value) {
+        this.value = value;
+    }
+
+    public ConfigChangeType getChangeType() {
+        return changeType;
+    }
+
+    public void setChangeType(ConfigChangeType changeType) {
+        this.changeType = changeType;
+    }
+}
